icount-api/requirements.txt
# python -m venv .venv
# or
# python3 -m venv .venv


### On Windows (cmd)
### Run project in venv
# .venv\Scripts\activate.bat
# .venv\Scripts\python.exe -m pip install --upgrade pip
# .venv\Scripts\python.exe -m pip install --upgrade -r requirements.txt

# pip freeze > requirements.txt

### On MacOS
### Run project in venv
# source .venv/bin/activate
# pip3 install --upgrade pip
# pip3 install --upgrade -r requirements.txt

# pip3 freeze > requirements.txt


### Packages
bcrypt==4.0.1
fastapi
pandas
passlib[bcrypt]
pydantic-settings
pyodbc
python-jose[cryptography]
sqlalchemy
uvicorn[standard]
